# docs/swagger.yaml
definitions:
  models.Hamburguer:
    properties:
      descricao:
        type: string
      id:
        type: integer
      ingredientes:
        items:
          $ref: '#/definitions/models.HamburguerIngrediente'
        type: array
      preco:
        type: number
    required:
    - descricao
    - preco
    type: object
  models.HamburguerIngrediente:
    properties:
      hamburguerID:
        type: integer
      item:
        $ref: '#/definitions/models.Item'
      itemID:
        type: integer
      quantidade:
        type: integer
    type: object
...
  models.Item:
    properties:
      descricao:
        type: string
      extra:
        description: true para "Com açúcar" em bebidas ou "Adicional" em ingredientes
        type: boolean
      id:
        type: integer
      preco:
        type: number
      tipo:
        $ref: '#/definitions/models.TipoItem'
    type: object
...
  models.PedidoBebida:
    properties:
      bebida:
        $ref: '#/definitions/models.Item'
      itemID:
        type: integer
      pedidoID:
        type: string
      quantidade:
        type: integer
    type: object
  models.PedidoHamburguer:
    properties:
      hamburguer:
        $ref: '#/definitions/models.Hamburguer'
      hamburguerID:
        type: integer
      pedidoID:
        type: string
      quantidade:
        type: integer
    type: object
  models.PedidoItemRequest:
    properties:
      id:
        type: integer
      quantidade:
        minimum: 1
        type: integer
    required:
    - id
    - quantidade
    type: object
  models.PedidoRequest:
    properties:
      bebidas:
        items:
          $ref: '#/definitions/models.PedidoItemRequest'
        type: array
      descricao:
        type: string
      endereco:
        type: string
      hamburgueres:
        items:
          $ref: '#/definitions/models.PedidoItemRequest'
        minItems: 1
        type: array
      nome:
        type: string
      observacoes:
        type: string
      telefone:
        type: string
    required:
    - descricao
    - endereco
    - hamburgueres
    - nome
    - telefone
    type: object
  models.PedidoResponse:
    properties:
      bebidas:
        items:
          $ref: '#/definitions/models.PedidoBebida'
        type: array
      data:
        type: string
      descricao:
        type: string
      endereco:
        type: string
      hamburgueres:
        items:
          $ref: '#/definitions/models.PedidoHamburguer'
        type: array
      id:
        type: string
      nome:
        type: string
      observacoes:
        type: string
      status:
        $ref: '#/definitions/models.StatusPedido'
      telefone:
        type: string
      valor_total:
        type: number
    type: object
  models.PedidoUpdateRequest:
    properties:
      bebidas:
        items:
          $ref: '#/definitions/models.PedidoItemRequest'
        type: array
      descricao:
        type: string
      endereco:
        type: string
      hamburgueres:
        items:
          $ref: '#/definitions/models.PedidoItemRequest'
        type: array
      nome:
        type: string
      observacoes:
        type: string
      status:
        $ref: '#/definitions/models.StatusPedido'
      telefone:
        type: string
    type: object
  models.StatusPedido:
    enum:
    - STARTED
    - DELIVERY
    - FINALIZED
    type: string
    x-enum-varnames:
    - StatusStarted
    - StatusDelivery
    - StatusFinalized
  models.TipoItem:
    enum:
    - BEBIDA
    - INGREDIENTE
    type: string
    x-enum-varnames:
    - TipoBebida
    - TipoIngrediente
info:
  contact: {}
paths:
...
  /pedidos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os pedidos cadastrados, com opção de
        filtrar por status não finalizado
      parameters:
      - description: Filtrar por status não finalizado (true/false)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PedidoResponse'
            type: array
      summary: Lista todos os pedidos
      tags:
      - pedidos
    post:
      consumes:
      - application/json
      description: Cria um novo pedido com os dados fornecidos
      parameters:
      - description: Dados do Pedido
        in: body
        name: pedido
        required: true
        schema:
          $ref: '#/definitions/models.PedidoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PedidoResponse'
        "400":
          description: Erro na validação dos dados
          schema:
            type: string
      summary: Cria um novo pedido
      tags:
      - pedidos
  /pedidos/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um pedido existente
      parameters:
      - description: ID do Pedido
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Erro ao deletar pedido
          schema:
            type: string
        "404":
          description: Pedido não encontrado
          schema:
            type: string
      summary: Deleta um pedido
      tags:
      - pedidos
    get:
      consumes:
      - application/json
      description: Retorna um pedido específico baseado no ID
      parameters:
      - description: ID do Pedido
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PedidoResponse'
        "404":
          description: Pedido não encontrado
          schema:
            type: string
      summary: Busca um pedido por ID
      tags:
      - pedidos
    put:
      consumes:
      - application/json
      description: Atualiza um pedido existente com os dados fornecidos
      parameters:
      - description: ID do Pedido
        in: path
        name: id
        required: true
        type: string
      - description: Dados do Pedido
        in: body
        name: pedido
        required: true
        schema:
          $ref: '#/definitions/models.PedidoUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PedidoResponse'
        "400":
          description: Erro na validação dos dados
          schema:
            type: string
        "404":
          description: Pedido não encontrado
          schema:
            type: string
      summary: Atualiza um pedido existente
      tags:
      - pedidos
swagger: "2.0"